New generator structure; new/changed config values
