+++ .README.md
@@ -4,9 +4,7 @@ ${badge('nodei')}
 
 ${badge('npm')}
 
-Hexo plugin to generate a indexes from custom front matter variables. Suppose
-you have an author variable in your front matter, this plugin will generate
-files like /authors/author-name.html, listing each post by that author.
+Hexo plugin to generate a indexes from custom front matter variables. Suppose you have an author variable in your front matter, this plugin will generate files like **/authors/author-name.html**, listing each post by that author.
 
 ## Install
 
@@ -18,8 +16,23 @@ an example. would be included in `_config.yml`
 
 ```yaml
 indexAnything:
-  author: authors
+  templateIndex: index
+  templateKey: index
+  titleSeparator: " - "
+  variables:
+    author: authors
 ```
+* ``templateIndex``  
+Template (.ejs) be used for the overview page of all available indexes
+
+* ``templateKey``  
+Template (.ejs) be used for the index page (list of posts) of a variable
+
+* ``titleSeparator``  
+String for separating the variable name (index) and the variable value (key) on the index page (list of posts) ... for example: **Author - John Doe**
+
+* ``variables``  
+List of all Frontmatter variables for which an index is created
 
 ## Usage
